Update abs_function_module.py
Updating wavelength_to_velocity function to be able to take a specified wavelength value as the reference value. This will be optional as to not disturb other programs that utilize this function. The default reference wavelength will be the same average CIV wavelength value that was utilized before this update.